UI and Powerups Fixed/Added
added some additional powerups to varying degrees of success, added some UI elements like the players health and coin count based on the manager script, and some other fixes to between UI, Powerups, and the Player

+++ Assets/Scripts/BulletScript.cs
@@ -44,7 +44,7 @@ public class BulletScript : MonoBehaviour
         {
             if (targetTeam.Contains(ob.GetComponent<HealthScript>().team))
             {
-                damage = ob.GetComponent<HealthScript>().TakeDamage(damage);
+                damage = ob.GetComponent<HealthScript>().TakeDamage((int)damage);
                 if (damage <= 0)
                 {
                     Destroy(this.gameObject);
